icelake: use unsigned shift base in utf16_to_latin1 tail mask (#976)
After the 32-element main loop, end - buf is in [1, 31]. With an input
length such that end - buf reaches 31 (e.g. len == 31), the existing
expression evaluates 'int 1 << 31', which places a 1 in the sign bit of
a 32-bit int. That is implementation-defined behaviour in C++17 (and was
undefined before). The rest of the icelake source already builds these
AVX-512 masks from an unsigned base — see e.g.
icelake_ascii_validation.inl.cpp, icelake_convert_utf16_to_utf8.inl.cpp,
and include/simdutf/internal/isadetection.h.

Switch the two tail-mask computations to '1U <<' to bring them in line.
No functional change on supported compilers.

+++ src/icelake/icelake_convert_utf16_to_latin1.inl.cpp
@@ -27,7 +27,7 @@ size_t icelake_convert_utf16_to_latin1(const char16_t *buf, size_t len,
     buf += 32;
   }
   if (buf < end) {
-    uint32_t mask(uint32_t(1 << (end - buf)) - 1);
+    uint32_t mask((1U << (end - buf)) - 1);
     __m512i in = _mm512_maskz_loadu_epi16(mask, buf);
     if (big_endian) {
       in = _mm512_shuffle_epi8(in, byteflip);
@@ -79,7 +79,7 @@ icelake_convert_utf16_to_latin1_with_errors(const char16_t *buf, size_t len,
     buf += 32;
   }
   if (buf < end) {
-    uint32_t mask(uint32_t(1 << (end - buf)) - 1);
+    uint32_t mask((1U << (end - buf)) - 1);
     __m512i in = _mm512_maskz_loadu_epi16(mask, buf);
     if (big_endian) {
       in = _mm512_shuffle_epi8(in, byteflip);
